requirements.txt
openai==0.28.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
aiofiles==23.2.0
typing-extensions==4.11.0